+++ features/bsm/service/api/src/main/java/org/opennms/netmgt/bsm/service/BusinessServiceManager.java
@@ -51,7 +51,7 @@ public interface BusinessServiceManager extends NodeManager {
 
     List<BusinessService> findMatching(Criteria criteria);
 
-    int countMatching(Criteria criteria);
+    long countMatching(Criteria criteria);
 
     BusinessService createBusinessService();
