+++ features/vaadin-dashlets/dashlet-summary/src/main/java/org/opennms/features/vaadin/dashboard/dashlets/SummaryDashlet.java
@@ -28,16 +28,23 @@
 
 package org.opennms.features.vaadin.dashboard.dashlets;
 
-import com.vaadin.server.Page;
-import com.vaadin.server.Sizeable;
-import com.vaadin.server.ThemeResource;
-import com.vaadin.ui.*;
 import org.opennms.core.criteria.CriteriaBuilder;
 import org.opennms.core.criteria.restrictions.SqlRestriction.Type;
 import org.opennms.features.vaadin.dashboard.model.*;
 import org.opennms.netmgt.dao.api.AlarmDao;
 import org.opennms.netmgt.model.OnmsAlarm;
 import org.opennms.netmgt.model.OnmsSeverity;
+
+import com.vaadin.server.Page;
+import com.vaadin.server.Sizeable;
+import com.vaadin.server.ThemeResource;
+import com.vaadin.ui.Accordion;
+import com.vaadin.ui.Alignment;
+import com.vaadin.ui.Component;
+import com.vaadin.ui.HorizontalLayout;
+import com.vaadin.ui.Image;
+import com.vaadin.ui.Label;
+import com.vaadin.ui.VerticalLayout;
 
 /**
  * This class implements a {@link Dashlet} for testing purposes.
@@ -138,7 +145,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param notAck number of unacknowledged alarms
      * @return the trend value
      */
-    private int computeTrend(int ack, int notAck) {
+    private int computeTrend(long ack, long notAck) {
         if (ack == notAck) {
             return TREND_EAST;
         } else {
@@ -224,8 +231,8 @@ public class SummaryDashlet extends AbstractDashlet {
             horizontalLayout.addStyleName("summary");
             horizontalLayout.addStyleName(onmsSeverity.name().toLowerCase());
 
-            int acknowledged = countBySeverity(true, m_timeslot, onmsSeverity);
-            int notAcknowledged = countBySeverity(false, m_timeslot, onmsSeverity);
+            long acknowledged = countBySeverity(true, m_timeslot, onmsSeverity);
+            long notAcknowledged = countBySeverity(false, m_timeslot, onmsSeverity);
 
             Label labelSeverity = new Label(onmsSeverity.getLabel());
             labelSeverity.addStyleName("summary-font");
@@ -327,8 +334,8 @@ public class SummaryDashlet extends AbstractDashlet {
                 }
             }
 
-            int acknowledged = countByUei(true, m_timeslot, uei);
-            int notAcknowledged = countByUei(false, m_timeslot, uei);
+            long acknowledged = countByUei(true, m_timeslot, uei);
+            long notAcknowledged = countByUei(false, m_timeslot, uei);
 
             Label labelSeverity = new Label(uei.replace("uei.opennms.org/nodes/", ""));
             labelSeverity.addStyleName("summary-font");
@@ -407,7 +414,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param onmsSeverity the {@link OnmsSeverity} to search for
      * @return number of alarms found
      */
-    public int countBySeverity(boolean acknowledged, long age, OnmsSeverity onmsSeverity) {
+    public long countBySeverity(boolean acknowledged, long age, OnmsSeverity onmsSeverity) {
         CriteriaBuilder criteriaBuilder = new CriteriaBuilder(OnmsAlarm.class);
 
         if (acknowledged) {
@@ -431,7 +438,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param uei          search for alarms with the specified uei
      * @return number of alarms found
      */
-    public int countByUei(boolean acknowledged, long age, String uei) {
+    public long countByUei(boolean acknowledged, long age, String uei) {
         CriteriaBuilder criteriaBuilder = new CriteriaBuilder(OnmsAlarm.class);
 
         if (acknowledged) {
