+++ opennms-tools/access-point-monitor/src/main/java/org/opennms/netmgt/accesspointmonitor/rest/AccessPointRestService.java
@@ -107,7 +107,7 @@ public class AccessPointRestService {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_accessPointDao.countAll());
+            return Long.toString(m_accessPointDao.countAll());
         } finally {
             readUnlock();
         }
