+++ opennms-web-api/src/main/java/org/opennms/web/svclayer/DistributedStatusService.java
@@ -82,5 +82,5 @@ public interface DistributedStatusService {
      *
      * @return a int.
      */
-    public int getApplicationCount();
+    public long getApplicationCount();
 }
