+++ opennms-web-api/src/main/java/org/opennms/web/svclayer/support/DefaultDistributedStatusService.java
@@ -199,7 +199,7 @@ public class DefaultDistributedStatusService implements DistributedStatusService
      * @return a int.
      */
     @Override
-    public int getApplicationCount() {
+    public long getApplicationCount() {
         return m_applicationDao.countAll();
     }
