+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/v1/AlarmRestService.java
@@ -108,7 +108,7 @@ public class AlarmRestService extends AlarmRestServiceBase {
     @Transactional
     public String getCount(@Context SecurityContext securityContext) {
         SecurityHelper.assertUserReadCredentials(securityContext);
-        return Integer.toString(m_alarmDao.countAll());
+        return Long.toString(m_alarmDao.countAll());
     }
 
     /**
