+++ opennms-webapp-rest/src/test/java/org/opennms/web/rest/v1/CategoryRestServiceIT.java
@@ -94,7 +94,7 @@ public class CategoryRestServiceIT extends AbstractSpringJerseyRestTestCase {
         String xml = sendRequest("GET", "/categories", 200);
         assertNotNull(xml);
         OnmsCategoryCollection categories = JAXB.unmarshal(new StringReader(xml), OnmsCategoryCollection.class);
-        int initialSize = categories.size();
+        long initialSize = categories.size();
         assertNotNull(categories);
         assertEquals(initialSize,categories.size());
