+++ opennms-webapp/src/test/java/org/opennms/web/outage/DaoWebOutageRepositoryIT.java
@@ -180,7 +180,7 @@ public class DaoWebOutageRepositoryIT implements InitializingBean {
     @Test
     @Transactional
     public void testCountMatchingOutages(){
-        int count = m_daoOutageRepo.countMatchingOutages(new OutageCriteria());
+        long count = m_daoOutageRepo.countMatchingOutages(new OutageCriteria());
         assertEquals(3, count);
         
         count = m_daoOutageRepo.countMatchingOutages(new OutageCriteria(new RegainedServiceDateBeforeFilter(new Date())));
